Preserve window buffer history
The per-window buffer history (window-prev-buffers) was being updated
when previewing buffers with pc-bufsw, even though switch-to-buffer
was being called with NORECORD. This interfered with
pc-bufsw-prefer-current-window, as it show the buffers in the order
they were last previewed, rather than actually selected.

Fix this by ensuring window-prev-buffers/window-next-buffers are
preserved when previewing buffers, and then update the buffer
histories with only the original/target buffer.